# Conflicts:
#	HomeCleanAPI/.idea/misc.xml

+++ HomeCleanAPI/src/main/java/com/example/homecleanapi/models/Feedback.java
@@ -14,7 +14,7 @@ public class Feedback {
     @JoinColumn(name = "job_id", nullable = false)
     private Job job; // Liên kết với bảng job
 
-    private Integer rating; // Đánh giá (1-5)
+    private Double rating; // Đánh giá (1-5)
     private String comment;
 
     // Getters and Setters
@@ -34,11 +34,11 @@ public class Feedback {
         this.job = job;
     }
 
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
